feat(canvas): in-place event editor on active card (extract CalendarEventForm)
