+++ apps/desktop/src/renderer/src/pages/canvas/canvas-event-editor.tsx
@@ -1,0 +1,116 @@
+/**
+ * CanvasEventEditor — hosts the extracted <CalendarEventForm> for an active
+ * event card on the spatial canvas (M6 Task 7, matrix #22 event). Owns the
+ * CalendarEventDraft; seeds it from calendarService.getEvent (ISO times
+ * converted to the form's local wall-clock strings, mirroring
+ * pages/calendar.tsx's createDraftFromItem) and saves via
+ * calendarService.updateEvent (draft converted back to ISO via the same
+ * localInputToIso helper the calendar page uses in toCreatePayload).
+ */
+import React, { useEffect, useState } from 'react'
+import { useT } from '@memry/i18n/renderer'
+import { CalendarEventForm } from '@/components/calendar/calendar-event-form'
+import {
+  localInputToIso,
+  toLocalDateInputValue,
+  toLocalDateTimeInputValue
+} from '@/components/calendar/date-utils'
+import type { CalendarEventDraft } from '@/components/calendar/types'
+import { calendarService, type CalendarEventRecord } from '@/services/calendar-service'
+import { extractErrorMessage } from '@/lib/ipc-error'
+import { createLogger } from '@/lib/logger'
+
+const log = createLogger('CanvasEventEditor')
+
+interface CanvasEventEditorProps {
+  eventId: string
+  onDone: () => void
+}
+
+function toDraft(event: CalendarEventRecord): CalendarEventDraft {
+  return {
+    title: event.title,
+    description: event.description ?? '',
+    isAllDay: event.isAllDay,
+    startAt: event.isAllDay
+      ? toLocalDateInputValue(event.startAt)
+      : toLocalDateTimeInputValue(event.startAt),
+    endAt: event.endAt
+      ? event.isAllDay
+        ? toLocalDateInputValue(event.endAt)
+        : toLocalDateTimeInputValue(event.endAt)
+      : '',
+    targetCalendarId: event.targetCalendarId
+  }
+}
+
+export const CanvasEventEditor = ({
+  eventId,
+  onDone
+}: CanvasEventEditorProps): React.JSX.Element => {
+  const [draft, setDraft] = useState<CalendarEventDraft | null>(null)
+  const [isSaving, setSaving] = useState(false)
+  const { t: tCommon } = useT('common')
+
+  useEffect(() => {
+    let cancelled = false
+    calendarService
+      .getEvent(eventId)
+      .then((event) => {
+        if (cancelled || !event) return
+        setDraft(toDraft(event))
+      })
+      .catch((error: unknown) => {
+        log.error('Failed to load calendar event', {
+          eventId,
+          error: extractErrorMessage(error)
+        })
+      })
+    return () => {
+      cancelled = true
+    }
+  }, [eventId])
+
+  if (!draft) {
+    return (
+      <div className="min-h-0 flex-1 overflow-auto p-3 text-[13px] text-text-tertiary">
+        {tCommon('state.loading')}
+      </div>
+    )
+  }
+
+  const onSave = async (): Promise<void> => {
+    setSaving(true)
+    try {
+      const result = await calendarService.updateEvent({
+        id: eventId,
+        title: draft.title.trim(),
+        description: draft.description.trim() || null,
+        startAt: localInputToIso(draft.startAt, draft.isAllDay),
+        endAt: draft.endAt ? localInputToIso(draft.endAt, draft.isAllDay) : null,
+        timezone: Intl.DateTimeFormat().resolvedOptions().timeZone || 'UTC',
+        isAllDay: draft.isAllDay,
+        targetCalendarId: draft.targetCalendarId
+      })
+      if (!result.success) {
+        throw new Error(result.error ?? 'Could not save event.')
+      }
+      onDone()
+    } finally {
+      setSaving(false)
+    }
+  }
+
+  return (
+    <div className="min-h-0 flex-1 overflow-auto p-3">
+      <CalendarEventForm
+        mode="edit"
+        draft={draft}
+        isSaving={isSaving}
+        onDraftChange={setDraft}
+        onSave={onSave}
+        onDismiss={onDone}
+      />
+    </div>
+  )
+}
